flask-server/requirements.txt
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
supabase==1.2.0
gunicorn==21.2.0
google-genai==1.25.0